Allow substitutions in the input file
refs #5649
