+++ test/tests/parser/param_substitution/param_substitution_in_file.i
@@ -1,0 +1,56 @@
+# Here we define a global parameter to be used for substitutions within this file
+FILENAME = 'special_string'
+
+[Mesh]
+  type = GeneratedMesh
+  dim = 2
+  nx = 10
+  ny = 10
+[]
+
+[Variables]
+  [./u]
+  [../]
+[]
+
+[Kernels]
+  [./diff]
+    type = Diffusion
+    variable = u
+  [../]
+[]
+
+[BCs]
+  [./left]
+    type = DirichletBC
+    variable = u
+    boundary = left
+    value = 0
+  [../]
+  [./right]
+    type = DirichletBC
+    variable = u
+    boundary = right
+    value = 1
+  [../]
+[]
+
+[Executioner]
+  type = Steady
+
+  # Preconditioned JFNK (default)
+  solve_type = 'PJFNK'
+
+
+  petsc_options_iname = '-pc_type -pc_hypre_type'
+  petsc_options_value = 'hypre boomeramg'
+[]
+
+[Outputs]
+  exodus = true
+
+  # Here we use the GetPot "DBE" function to perform a substitution.
+  # The parameter "FILENAME" can either exist in this file or
+  # be provided on the CLI
+  file_base = out_${FILENAME}
+[]
