+++ test/tests/parser/param_substitution/tests
@@ -1,0 +1,18 @@
+[Tests]
+  [./param_substitution_cli]
+    type = 'Exodiff'
+    input = 'param_substitution.i'
+
+    # Here we provide the parameter we want to sub on the Command line
+    cli_args = 'FILENAME=special_string'
+    exodiff = 'out_special_string.e'
+  [../]
+
+  [./param_substitution_in_file]
+    type = 'Exodiff'
+    input = 'param_substitution_in_file.i'
+
+    exodiff = 'out_special_string.e'
+    prereq = 'param_substitution_cli'
+  [../]
+[]
